+++ .buildkite/ftr_platform_stateful_configs.yml
@@ -265,6 +265,7 @@ enabled:
   - x-pack/test/functional/apps/rollup_job/config.ts
   - x-pack/test/functional/apps/saved_objects_management/config.ts
   - x-pack/test/functional/apps/saved_query_management/config.ts
+  - x-pack/test/functional/apps/saved_query_management/config.v2.ts
   - x-pack/test/functional/apps/security/config.ts
   - x-pack/test/functional/apps/snapshot_restore/config.ts
   - x-pack/test/functional/apps/spaces/config.ts
