+++ x-pack/test/functional/apps/saved_query_management/config.v2.ts
@@ -12,6 +12,6 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
 
   return {
     ...functionalConfig.getAll(),
-    testFiles: [require.resolve('./feature_controls/v1')],
+    testFiles: [require.resolve('./feature_controls/v2')],
   };
 }
