+++ x-pack/test/functional/apps/saved_query_management/feature_controls/v1/index.ts
@@ -5,11 +5,10 @@
  * 2.0.
  */
 
-import { FtrProviderContext } from '../../../ftr_provider_context';
+import { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ loadTestFile }: FtrProviderContext) {
-  describe('Feature controls', function () {
-    this.tags('skipFirefox');
+  describe('Saved query management - Feature controls', function () {
     loadTestFile(require.resolve('./security'));
   });
 }
