+++ x-pack/test/functional/apps/saved_query_management/feature_controls/v2/index.ts
@@ -5,10 +5,10 @@
  * 2.0.
  */
 
-import { FtrProviderContext } from '../../ftr_provider_context';
+import { FtrProviderContext } from '../../../../ftr_provider_context';
 
 export default function ({ loadTestFile }: FtrProviderContext) {
-  describe('Saved query management', function () {
-    loadTestFile(require.resolve('./feature_controls'));
+  describe('Saved query management - Feature controls - v2', function () {
+    loadTestFile(require.resolve('./security.v2.ts'));
   });
 }
